fix(windows): add missing Flutter method channel includes
- Add flutter/method_channel.h, standard_method_codec.h, and
  encodable_value.h includes to flutter_window.cpp
- Cast size_t to DWORD in auto_start.cpp to fix warning

+++ flutter_app/windows/runner/auto_start.cpp
@@ -23,7 +23,7 @@ extern "C" __declspec(dllexport) bool EnableAutoStart(const char* appPath) {
         0,
         REG_SZ,
         (const BYTE*)value.c_str(),
-        value.length() + 1
+        static_cast<DWORD>(value.length() + 1)
     );
 
     RegCloseKey(hKey);
